PhaseFrac and PhaseFrac1 use uint32 instead of long argument
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@9391 a380766d-ff14-0410-b294-a243070f3f08

+++ Headers/plugin_interface/SC_SndBuf.h
@@ -50,16 +50,16 @@ typedef struct SndBufUpdates SndBufUpdates;
 enum { coord_None, coord_Complex, coord_Polar };
 
 
-inline float PhaseFrac(unsigned long inPhase) 
+inline float PhaseFrac(uint32 inPhase)
 		{
-			union { unsigned long itemp; float ftemp; } u;
+			union { uint32 itemp; float ftemp; } u;
 			u.itemp = 0x3F800000 | (0x007FFF80 & ((inPhase)<<7));
 			return u.ftemp - 1.f;
 		}
 		
-inline float PhaseFrac1(unsigned long inPhase) 
+inline float PhaseFrac1(uint32 inPhase)
 		{
-			union { unsigned long itemp; float ftemp; } u;
+			union { uint32 itemp; float ftemp; } u;
 			u.itemp = 0x3F800000 | (0x007FFF80 & ((inPhase)<<7));
 			return u.ftemp;
 		}
